feat: CI gates, canonical fixtures, checkpoint waivers, and status audit
CI infrastructure (bd-15k, bd-16r):
- Add tools/ci/validate_schemas.sh — JSON schema validation gate using
  jsonschema (Draft 2020-12 with Draft7 fallback) that validates fixture
  files against schemas/, emits conformance artifacts, and supports
  --strict mode. Wire as lint-schema-validation Makefile target in both
  check and check-ci combined gates.
- Add .github/workflows/nightly.yml — daily cron workflow with fuzz-nightly
  (100K iterations, 60-min timeout, minimizer self-test) and e2e-extended
  (vertical lanes, conformance, codec equivalence, profile parity, schema
  validation in --strict mode) jobs with 14-day artifact retention.

Canonical fixture families:
- Add 44 Rust-reference fixture files across 5 new directories under
  fixtures/rust_reference/ covering core_budget (meet-identity, exhaustion-
  rollback), core_cancel (user-cancel, severity-strengthen, timeout-budget),
  core_channel (two-phase, fifo-ordering, backpressure), core_lifecycle
  (region open/close/drain/finalize, task spawn/complete, obligation
  reserve/commit/abort, finalization-quiescence), and core_timer (register-
  fire, equal-deadline-ordering, stale-cancel-rejected). Each includes
  codec_bin.json and codec_json.json variants per applicable profile.

Source and tests:
- Add ASX_CHECKPOINT_WAIVER annotations to 7 bounded loops across mpsc.c,
  mpsc_lockfree_spike.c, cleanup.c, and quiescence.c to satisfy the
  checkpoint-coverage lint gate.
- Add doc comments to asx_abi_version_{major,minor,patch} in asx_abi.h.
- Cast m.cost_quota to uint32_t in test_wasm32_oracle.c (fix -Wconversion).
- Implement snapshot determinism vignette in vignette_replay.c: runs
  scenario twice, captures asx_snapshot_buffer, computes digest, and
  asserts both runs match.
- Add Makefile build rules for test_arena_locality, test_barrier_cert,
  and test_seqlock_ebr spike binaries.

Documentation status audit:
- Advance all 21 semantic units to impl-complete in FEATURE_PARITY.md
  based on 59 test suites, invariant tests, CBMC/algebraic/litmus gates,
  14 e2e fixture families (85+ scenarios), and clean -Werror build.
- Mark TRC-INV-001, TRC-GSM-001/002/003, TRC-BUILD-001 as complete in
  PLAN_EXECUTION_TRACEABILITY_INDEX.md.
- Update QUALITY_GATES.md: SLO-based blocking now active via
  evaluate_slo_gates.sh --strict in perf.yml.
- Replace "pending" commit placeholders with actual SHA in
  size_cold_start_baselines.json and slo_baselines.json.

Beads: close bd-15k (nightly workflow), bd-16r (schema validation),
bd-1cg (traceability audit).

+++ tests/unit/runtime/test_wasm32_oracle.c
@@ -156,7 +156,7 @@ TEST(budget_meet_digest_is_stable)
 
     h = fnv1a_u32(h, (uint32_t)m.deadline);
     h = fnv1a_u32(h, m.poll_quota);
-    h = fnv1a_u32(h, m.cost_quota);
+    h = fnv1a_u32(h, (uint32_t)m.cost_quota);
     h = fnv1a_u32(h, m.priority);
 
     /* Identity: meet with infinite */
@@ -166,7 +166,7 @@ TEST(budget_meet_digest_is_stable)
 
     h = fnv1a_u32(h, (uint32_t)m.deadline);
     h = fnv1a_u32(h, m.poll_quota);
-    h = fnv1a_u32(h, m.cost_quota);
+    h = fnv1a_u32(h, (uint32_t)m.cost_quota);
     h = fnv1a_u32(h, m.priority);
 
     ASSERT_TRUE(h != 0);
